requirements.txt
pydantic~=1.10.2
requests~=2.28.1
prettytable~=3.3.0
colorama~=0.4.5
PyYAML~=6.0
connexion~=2.14.1
gevent~=22.10.2
six~=1.16.0
beautifulsoup4~=4.11.1